[ALPS05144979] ISP: fix code defect
[Detail] Fix code defect to prevent overflow.

MTK-Commit-Id: dd7050290e6239890201e9515079da2a0120b89c

CR-Id: ALPS05144979
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/cameraisp/src/mt6739/camera_isp.c
@@ -11699,7 +11699,7 @@ static signed int mmap_kmem(struct file *filp, struct vm_area_struct *vma)
 static signed int ISP_mmap(struct file *pFile, struct vm_area_struct *pVma)
 {
 	unsigned long length = 0;
-	unsigned int pfn = 0x0;
+	unsigned long pfn = 0x0;
 
 	LOG_DBG("- E.");
 	length = (pVma->vm_end - pVma->vm_start);
